testes folder in controle_estoque created with tests separated by file
